Adjusted how Demacro handles the new ignore parameter Fixed a bug in the number of args for a callable macro
Changed ignore_newcommand to ignore_new_macros.  The idea for this parameter is that when it's enabled, newcommand, newenvironment, etc. are ignored.  In the future the implementation of this will need to be adjusted.

I adjusted cli.py accordingly.  There was a small bug in this code which I fixed as well.